move image picker to components and bottom sheets to interaction listener package
